Move soft_u2f_v2 into installed package and update test imports
