+++ test/test_u2f_highlevel.py
@@ -26,7 +26,7 @@
 # POSSIBILITY OF SUCH DAMAGE.
 
 from u2flib_server import u2f
-from .soft_u2f_v2 import SoftU2FDevice
+from u2flib_server.soft_u2f_v2 import SoftU2FDevice
 import unittest
 
 APP_ID = 'https://www.example.com'
